test: 타입을 기준으로 테스트 케이스 파일 분리 #3
